==> MarkdownDocs/1_nightjar_survey_maps.Rmd <==
---
title: "1_nightjar_survey_maps"
author: "Marie I. Tosa"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#load packages
require(tidyr)
require(plyr)

require(sf)
require(raster)

require(ggplot2)

```

## nightjar survey maps from nightjar survey network and non nightjar survey network
```{r echo=F}
```

##load GIS layers
```{r, echo=T}
#these are in NAD 83, longlat
AF_states <- st_read(dsn="../../GIS", layer="AF_states") 
# BCR <- read_sf(dsn="../../GIS", layer="BCR_AF_clip") #this one is outdated
BCR <- read_sf(dsn="../../GIS", layer="BCR-state") #correct one downloaded 7/1/2023
```

##load data
```{r echo=T}
master.nsn <- read.table(file="../Data_processed/master_nsn_2023-07-24.txt", sep=",", header=T) #fixed multiple rows per survey issue, 8/17/2023
# master.nonnsn <- read.table(file="../Data_processed/master_nonnsn_2023-06_21.txt", sep=",", header=T)
master.nonnsn <- read.table(file="../Data_processed/master_nonnsn_2023-08-17.txt", sep=",", header=T) #added MA, 8/17/2023

table(master.nsn$state)
table(master.nonnsn$state)
```

##create file of all survey locations
```{r echo=T}
#route locations for NSN surveys since no stop information in master.nsn file
rt <- read.csv(file="../Data_raw/Copy of NSN Data/NSN Data (March 28 2023) - Routes.csv")
#duplicate routeid when there are stops 1-11
#Latitude and Longitude are flipped

  #convert 5 digit route id to 6 digit
  rt$Routeid.char <- sprintf("%06i", rt$Routeid)
  rt$st_code <- substr(rt$Routeid.char, 0, 2)
  rt$rt_code <- substr(rt$Routeid.char, 3, 6)
  
  #add state letters
  state.codes <- read.csv("../Data_raw/NSN_routeid_codes.csv", colClasses = "character")
  rt <- merge(rt, state.codes, by.x="st_code", by.y="routeid_code", all.x=T)
  rt <- rt[rt$state %in% AF_states$ST,]
  rt$st.route <- paste(rt$state, rt$rt_code, sep="_")
  # table(rt$st.route)
  
  rt <- rt[,c("Routeid.char","Stop","state","Latitude","Longitude","st.route")]
  names(rt) <- c("Routeid.char","Stop","state","Longitude","Latitude","st.route")

rt2 <- unique(master.nonnsn[,c("Route.Stop","state","name","Route.Latitude","Route.Longitude")])
rt2$st.route <- paste(rt2$state, rt2$name, sep="_")
rt2$id <- rt2$name
# table(rt2[,"st.route"])
rt2 <- rt2[,c("id","Route.Stop","state","Route.Latitude","Route.Longitude","st.route")]
names(rt2) <- c("Routeid.char","Stop","state","Latitude","Longitude","st.route")

rt.all <- rbind(rt, rt2)
rt.all <- rt.all[!duplicated(rt.all[,c("st.route","Stop")]),]
write.table(rt.all, file="../Data_processed/master_survey_locations.txt", sep=",", row.names=F) #save survey locations

table(rt.all$state, useNA = "always")
#check FL routes
table(rt.all[rt.all$state == "FL",]$st.route) #a bunch of routes with only 1 stop?

#routes missing Lat/Long data
nrow(rt.all[is.na(rt.all$Latitude),]) #18 rows
#NY_Rocky Point, NY_LOSP1, NY_Fishers Island, NY_Mayfield, NY_, NH_3 rivers Woods, NH_, NH_Lyndeborough, VT_PittsfordVT
rt.all.sf <- st_as_sf(rt.all[!is.na(rt.all$Latitude),], coords=c("Longitude","Latitude"), crs=crs(AF_states))

ggplot(data=rt.all.sf) + geom_sf(data=AF_states) + geom_sf(size=1) + theme_bw(base_size=20) + theme(panel.grid=element_blank())

```

##subset data: only include data within Atlantic flyway and data with location information
```{r, echo=T}

#nsn data
nsn <- master.nsn[!is.na(master.nsn$Latitude),]
na.nsn <- master.nsn[is.na(master.nsn$Latitude),]

nrow(master.nsn) #52,210 rows 8/23/2023; 71,307 rows #number of observations 
nrow(nsn)        #17,235 rows 8/23/2023; 44,503 rows #number of observations with Lat/Long info

af.nsn <- master.nsn[master.nsn$state %in% AF_states$ST,]
af.nsn$species <- af.nsn$code
af.nsn[is.na(af.nsn$species),]$species <- "NONE"
af.nsn.sf <- st_as_sf(af.nsn[!is.na(af.nsn$Latitude),],
                      coords=c("Latitude","Longitude"),
                      crs=4269)

nrow(af.nsn)    #19510 rows
nrow(af.nsn.sf) #6016 rows

#non nsn data
nrow(master.nonnsn)                                        #now 14,486 rows 8/18/2023
nrow(master.nonnsn[!is.na(master.nonnsn$Route.Latitude),]) #now 14,396 rows

nonnsn.sf <- st_as_sf(master.nonnsn[!is.na(master.nonnsn$Route.Latitude),], coords=c("Route.Longitude","Route.Latitude"), crs=4269)

#combine nsn and non-nsn survey data
nsn.locs <- af.nsn[,c("Latitude","Longitude","start_year","route_id.char","state")]
  names(nsn.locs) <- c("Longitude","Latitude","year","route_id.char","state")
  nsn.locs$Route.Stop <- 1
nonnsn.locs <- master.nonnsn[,c("Route.Longitude","Route.Latitude","year","id","state","Route.Stop")]
  names(nonnsn.locs) <- c("Longitude","Latitude","year","route_id.char","state","Route.Stop")


all.locs <- rbind(nsn.locs, nonnsn.locs)
all.locs.sf <- st_as_sf(all.locs[!is.na(all.locs$Longitude),], 
                        coords=c("Longitude","Latitude"),
                        crs=4269)

table(all.locs[,c("state","year")])

all.locs$year <- as.numeric(all.locs$year)
surveys.year.long <- ddply(all.locs, .(state, year), 
                           summarize, numsurveys=length(route_id.char),
                           numroutes=length(unique(route_id.char)))
surveys.year.long <- surveys.year.long[order(surveys.year.long$year),]
surveys.year.wide <- pivot_wider(surveys.year.long, 
                                 names_from="year", id_cols="state",
                                 values_from="numsurveys", values_fill = 0)
surveys.year.wide
write.csv(surveys.year.wide,
          file="../Data_processed/summary_nsn_non-nsn_surveys_perYear_perState.csv")

routes.year.wide <- pivot_wider(surveys.year.long, 
                                 names_from="year", id_cols="state",
                                 values_from="numroutes", values_fill = 0)
routes.year.wide
write.csv(routes.year.wide,
          file="../Data_processed/summary_nsn_non-nsn_routes_perYear_perState.csv")

#current surveys (2019-2023)
current.surveys <- ddply(all.locs[all.locs$year %in% c(2019:2023),], .(state),
                         summarize, numcurrentsurveys=length(route_id.char),
                         numcurrentroutes=length(unique(route_id.char)))
current.surveys
write.csv(current.surveys,
          file="../Data_processed/summary_nsn_non-nsn_surveys_current.csv")

```
## plot observation data
```{r, echo=T}

base <- ggplot() + geom_sf(data=AF_states, fill=NA) + 
  theme_bw(base_size = 20) + theme(panel.grid=element_blank())

p.all <- base + 
  geom_sf(data=af.nsn.sf, col="steelblue") + 
  geom_sf(data=nonnsn.sf, col="navy")
p.all
ggsave(p.all, filename="../Figures/NSN/all_surveys.tiff", height=10, width=8,
       units="in", dpi=400, compression="lzw")

base2 <- base + 
  geom_sf(data=all.locs.sf, pch=1, col="black", size=0.5, stroke=0.1)
  
unique(af.nsn$code)
for(sp in unique(af.nsn$code)[1:5])
{
  sp.map <- base2 + 
          geom_sf(data=af.nsn.sf[af.nsn.sf$code %in% sp,],
                  aes(col=code), size=2) + 
          geom_sf(data=nonnsn.sf[nonnsn.sf$species %in% sp,],
                  aes(col=species), size=2)
  ggsave(sp.map,
         filename=paste("../Figures/NSN/all_surveys_", sp, ".tiff", sep=""),
         height=10, width=8, units="in", dpi=400, compression="lzw")
}

# sp <- c("ANNI","CONI","CWWI","EWPW","NONE")
sp <- c("CONI","CWWI","EWPW","NONE")
sp <- c("CONI","CWWI","EWPW")

sp.map <- base2 + 
          geom_sf(data=af.nsn.sf[af.nsn.sf$species %in% sp,], 
          # geom_sf(data=af.nsn.sf[!af.nsn.sf$species %in% c("NONE","NA"),], 
                  aes(col=species), size=1) + 
          geom_sf(data=nonnsn.sf[nonnsn.sf$species %in% sp,],
          # geom_sf(data=nonnsn.sf[!nonnsn.sf$species %in% c("NONE","NA"),], 
                  aes(col=species), size=1) +
    facet_wrap(~species, nrow=1) + theme(legend.position = "none")

# ggsave(sp.map, filename="../Figures/NSN/all_surveys_sp.tiff", height=8, width=20,
#        units="in", dpi=400, compression="lzw") #for maps with ANNI
# 
# ggsave(sp.map, filename="../Figures/NSN/all_surveys_sp2.tiff", height=8, width=20, #no ANNI maps
ggsave(sp.map, filename="../Figures/NSN/all_surveys_sp3.tiff", height=8, width=20, #no ANNI or NONE maps
       units="in", dpi=400, compression="lzw") 

for(y in 2008:2022)
{
  print(y)
  sp.map <- base + 
    geom_sf(data=all.locs.sf[all.locs.sf$year %in% y,],
          pch=1, col="black", size=0.5, stroke=0.1) +
    geom_sf(data=af.nsn.sf[af.nsn.sf$start_year %in% y & af.nsn.sf$species %in% sp,], 
                  aes(col=species), size=1) + 
    geom_sf(data=nonnsn.sf[nonnsn.sf$year %in% y & nonnsn.sf$species %in% sp,], 
                  aes(col=species), size=1) +
    facet_wrap(~species, nrow=1) + theme(legend.position = "none")

ggsave(sp.map, filename=paste("../Figures/NSN/surveys_by_year/all_surveys_sp_", y, ".tiff", sep=""),
       height=8, width=16, units="in", dpi=400, compression="lzw")
}

```

## plot current routes (2019-present)

```{r, echo=T}
#include all stops on routes
y <- 2019:2023
rt.sf <- st_as_sf(rt, coords=c("Latitude","Longitude"), crs=4269)

#subsetted below, so not necessary
# rt.current.sf <- rt.sf[rt.sf$Routeid.char %in% 
#                          unique(all.locs[all.locs$year %in% y,]$route_id.char),]

sp.map <- base + 
    # geom_sf(data=all.locs.sf[all.locs.sf$year %in% y,],
    #       pch=1, col="black", size=0.5, stroke=0.1) +
    # geom_sf(data=rt.sf[rt.sf$Routeid.char %in% 
    #                      unique(all.locs[all.locs$year %in% y,]$route_id.char),],
    #         pch=1, col="black", size=0.5, stroke=0.1) +
    geom_sf(data=rt.all.sf, pch=1, col="black", size=0.5, stroke=0.1) +
    geom_sf(data=af.nsn.sf[af.nsn.sf$start_year %in% y & af.nsn.sf$code %in% sp,], 
                  aes(col=species), size=1) + 
    geom_sf(data=nonnsn.sf[nonnsn.sf$year %in% y & nonnsn.sf$species %in% sp,], 
                  aes(col=species), size=1) +
    facet_wrap(~species, nrow=1) + theme(legend.position = "none")
sp.map

ggsave(sp.map, filename=paste("../Figures/NSN/2019-2023_surveys_sp.tiff",
                              sep=""),
       height=8, width=20, units="in", dpi=400, compression="lzw")
```

## plot each state separately
```{r, echo=T}
y <- 2019:2023
nonnsn.sf$year <- factor(nonnsn.sf$year) #convert to factor, so not continuous colors

for(s in AF_states$ST)
{
  if(s %in% c("DC","NJ","RI","VT"))
  {
    next
  }
  print(s)
  #state outline
  ex <- extent(AF_states[AF_states$ST == s,])
  base.st <-  ggplot() + 
    geom_sf(data=AF_states, fill=NA) + 
    theme_bw(base_size = 20) + 
    theme(panel.grid=element_blank(), legend.position = "none")
  
  #maps of all survey locations
  p.all <- base.st +
    geom_sf(data=rt.all.sf, 
            pch=1, col="grey50", size=0.5, stroke=0.1) +
    geom_sf(data=af.nsn.sf[af.nsn.sf$state == s,], aes(col=start_year),
            pch=1, size=0.5, stroke=1, fill=NA) + 
    geom_sf(data=rt.sf[rt.sf$Routeid.char %in% 
                         unique(all.locs[all.locs$year %in% y,]$route_id.char),],
            col="grey50", pch=1, size=0.5, stroke=1, fill=NA) +
    geom_sf(data=nonnsn.sf[nonnsn.sf$state == s,], aes(col=year),
            pch=1, size=0.5, stroke=1, fill=NA) +
    theme(legend.position="top") +
    coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax))
  # print(p.all)
  ggsave(p.all, filename=paste("../Figures/NSN/surveys_by_state/", s, "_all_surveys.tiff", sep=""),
       height=10, width=8, units="in", dpi=400, compression="lzw")
  
  #maps of current survey locations and detections of species
  p.sp <- base.st +
    geom_sf(data=all.locs.sf[all.locs.sf$year %in% y,],
          pch=1, col="black", size=0.5, stroke=0.1) +
    geom_sf(data=rt.sf[rt.sf$Routeid.char %in% 
                         unique(all.locs[all.locs$year %in% y,]$route_id.char),],
            pch=1, col="black", size=0.5, stroke=0.1) +
    geom_sf(data=af.nsn.sf[af.nsn.sf$start_year %in% y & af.nsn.sf$code %in% sp,], 
                  aes(col=species), size=2) + 
    geom_sf(data=nonnsn.sf[nonnsn.sf$year %in% y & nonnsn.sf$species %in% sp,], 
                  aes(col=species), size=2) +
    facet_wrap(~species, nrow=1) +
    coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax))
  
  ggsave(p.sp,
      filename=paste("../Figures/NSN/surveys_by_state/",s,"_2019-2023_surveys_sp.tiff", sep=""),
       height=8, width=20, units="in", dpi=400, compression="lzw")
}

```

## summarize data based on state and BCR information

### start here and check
```{r, echo=T}
#assign each point to BCR value
# unique.locs <- unique(all.locs[,c("route_id.char","state")])
unique.locs <- all.locs[!is.na(all.locs$Longitude),]
unique.locs <- unique.locs[!duplicated(unique.locs$route_id.char),
                        c("route_id.char","state","Longitude","Latitude")]
unique.locs.sf <- st_as_sf(unique.locs[!is.na(unique.locs$Longitude),], 
                           coords=c("Longitude","Latitude"),
                           crs=4269)
#assign BCRs and states to survey points
loc.bcr <- sapply(st_intersects(unique.locs.sf, BCR),
                  function(z) if (length(z)==0) NA_integer_ else z)
BCR.st <- BCR
st_geometry(BCR.st) = NULL
BCR.st <- as.data.frame(BCR.st)
BCR.st$bcr.id <- row.names(BCR.st)

loc.bcr <- data.frame(loc.bcr)
loc.bcr <- cbind(unique.locs.sf, loc.bcr)
loc.bcr <- merge(loc.bcr, BCR.st, by.x="loc.bcr", by.y="bcr.id", all.x=T)
loc.bcr$BCR <- factor(loc.bcr$BCR)

loc.bcr.na <- loc.bcr[is.na(loc.bcr$loc.bcr),]
nrow(loc.bcr.na)

loc.bcr.na[loc.bcr.na$state %in% "MA",]

###########
#fix loc.bcr that don't have a BCR value
loc.bcr[loc.bcr$state %in% "MD" & is.na(loc.bcr$loc.bcr),]$BCRNAME <- toupper("New England/Mid-Atlantic Coast")
loc.bcr[loc.bcr$state %in% "MD" & is.na(loc.bcr$loc.bcr),]$BCR <- 30

loc.bcr[loc.bcr$route_id.char %in% c("250009","250052","259336","250921","259242"),]$BCRNAME <- toupper("Southeastern Coastal Plain") #FL routes
loc.bcr[loc.bcr$route_id.char %in% c("250009","250052","259336","250921","259242"),]$BCR <- 27

loc.bcr[loc.bcr$state %in% "FL" & is.na(loc.bcr$BCRNAME),]$BCRNAME <- toupper("Peninsular Florida")
loc.bcr[loc.bcr$state %in% "FL" & is.na(loc.bcr$BCR),]$BCR <- 31

loc.bcr[loc.bcr$state %in% "VA" & is.na(loc.bcr$loc.bcr),]$BCRNAME <- toupper("New England/Mid-Atlantic Coast")
loc.bcr[loc.bcr$state %in% "VA" & is.na(loc.bcr$loc.bcr),]$BCR <- 30

loc.bcr[loc.bcr$state %in% "MA" & is.na(loc.bcr$loc.bcr),]$BCRNAME <- toupper("New England/Mid-Atlantic Coast")
loc.bcr[loc.bcr$state %in% "MA" & is.na(loc.bcr$loc.bcr),]$BCR <- 30

# loc.bcr[loc.bcr$state %in% "VT" & is.na(loc.bcr$loc.bcr),]$BCRNAME <- toupper("Atlantic Northern Forest")
# loc.bcr[loc.bcr$state %in% "VT" & is.na(loc.bcr$loc.bcr),]$BCR <- 14

# loc.bcr[loc.bcr$state %in% "NY" & is.na(loc.bcr$loc.bcr),]$BCRNAME <- toupper("New England/Mid-Atlantic Coast")
# loc.bcr[loc.bcr$state %in% "NY" & is.na(loc.bcr$loc.bcr),]$BCR <- 30

loc.bcr[is.na(loc.bcr$BCR),]
#remaining point says "NC", but coordinates are in Ontario

####
#save this since has st-bcr info
loc.bcr.df <- data.frame(loc.bcr)
write.csv(loc.bcr.df[,c("route_id.char","state","BCR","BCRNAME")], file="../Data_processed/route_bcr_state.csv", row.names=F)

############
BCR$BCR <- factor(BCR$BCR)
#check bcrs by plotting by color, does it match the bcr_labeled.tiff? Yes.
#need to correct BCRs of some points
nsn.plot <- ggplot() + 
  # geom_sf(data=BCR, aes(fill=BCRNumber), col="grey90") +
  # scale_fill_brewer(palette="Set2") +
  geom_sf(data=AF_states, col="grey75", fill=NA) +
  # geom_sf(data=loc.bcr.na, aes(col=BCR)) +
  geom_sf(data=loc.bcr, aes(col=BCR)) +
  scale_color_brewer(palette="Set2", na.value="black") +
  theme_bw(base_size=20) + 
  theme(panel.grid=element_blank(), legend.position=c(0.8, 0.25))
nsn.plot
ggsave(nsn.plot, filename = "../Figures/NSN/all_surveys_by_BCR-state.tiff",
       height=8, width=8, units="in", dpi=400, compression="lzw")

ex <- extent(AF_states[AF_states$ST == "NY",])
nsn.plot + coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax))

st_geometry(loc.bcr) = NULL #convert back to data frame

###########
#current surveys (2019-2023)
all.locs <- merge(all.locs, loc.bcr, by=c("route_id.char","state"), all.x=T) #33,996 obs.

#summarize by BCR-State
current.surveys <- ddply(all.locs[all.locs$year %in% c(2019:2023),],
                         .(state, BCR, BCRNAME),
                         summarize, numcurrentsurveys=length(route_id.char),
                         numcurrentroutes=length(unique(route_id.char)),
                         numyears=length(unique(year)))
current.surveys$meansurveysperyear <- round(current.surveys$numcurrentsurveys/current.surveys$numyears, digits=2)
write.csv(current.surveys,
          file="../Data_processed/summary_nsn_non-nsn_surveys_current_bcr-state.csv")
```

# create animation of detections of nightjars by year
```{r}
#make sure to run through 5_nsn_detectionhistories_2023-07-13.R before running this part
require(magick)

BCR$BCR <- factor(BCR$BCR)

#created this file in 5_nsn_detectionhistories_2023-07-13.Rmd
nightjars <- read.table("../Data_processed/master_alldata_2023-08-18.txt", sep=",", header=T) #22,339 observations
nightjars$BCR <- factor(nightjars$BCR)
nightjars.sf <- st_as_sf(nightjars[!is.na(nightjars$Route.Latitude),], coords=c("Route.Longitude","Route.Latitude"), crs=4269)

table(nightjars.sf$species)
# ANNI CONI CWWI EWPW NONE
#    2  763 2094 4272 9417

#create still images and the stitch together
sp <- c("CONI","CWWI","EWPW")
for(y in 2007:2022)
{
  print(y)
  ggplot() +
    geom_sf(data=BCR, aes(fill=BCR), alpha=0.3, col="grey80") + scale_fill_brewer(palette = "Set2") +
    geom_sf(data=nightjars.sf[nightjars.sf$year %in% y, c("year","site","BCR")], pch=1, col="grey25", size=1, stroke=0.1) +    
    geom_sf(data=nightjars.sf[nightjars.sf$year %in% y & nightjars.sf$species %in% sp,], aes(col=species, size=TotalDetections)) +
    facet_wrap(~species, nrow=1) + theme_bw(base_size=20) + theme(legend.position = "none", panel.grid=element_blank()) + 
    ggtitle(y)
  ggsave(filename=paste("../Figures/NSN/surveys_animation/a_", y, ".png", sep=""), height=8, width=16)
}

#animate with magick
list.files(path="../Figures/NSN/surveys_animation/", pattern=".png", full.names=T) %>%
  image_read() %>%
  image_join() %>%
  image_animate(fps=2) %>%
  image_write("../Figures/NSN/surveys_animation/NSN_surveys_2007-2022.gif")


```
