==> MarkdownDocs/2b_BBS_occupancy_unmarked_stacked_EWPW_2023-08-02.Rmd <==
---
title: "2b_BBS_stacked_single-season occupancy_EWPW"
author: "Marie I. Tosa"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#load packages

require(unmarked)
require(plyr)
require(dplyr)
require(tidyr)
require(chron) #for times function

require(ggplot2)
require(ggpubr)
```


## Create functions to plot figures
```{r}
vals.coefs <- function(mod)
{
coefs <- coef(mod) %>% tibble::enframe() #extract coefficients for covariates
se <- SE(mod) %>% tibble::enframe() #extract SEs for covariates

ci.psi <- confint(mod, type="state") %>% tibble::enframe()
ci.det <- confint(mod, type="det") %>% tibble::enframe()
ci <- rbind(ci.psi, ci.det) #combine psi and det CIs
ci <- data.frame(name=ci$name, lower=ci$value[,1], upper=ci$value[,2]) #need to split up ci so can read as data frame with lower and upper bounds

coefs <- merge(coefs, se, by="name") #merge coefficient values and ses
coefs <- merge(coefs, ci) #merge coefs with CIs
coefs <- data.frame(coefs)
names(coefs) <- c("par", "est", "se", "lower","upper")

coefs$p <- gsub(substr(coefs$par, 0,3), pattern="p\\([A-z]", replacement="det")
coefs$name <- gsub(coefs$par, pattern="scale\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="p\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="psi\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="\\)\\)", replacement="")
coefs$name <- gsub(coefs$name, pattern="Int\\)", replacement="Int")
coefs$name <- gsub(coefs$name, pattern="\\)", replacement="")
coefs$name <- gsub(coefs$name, pattern="I\\(", replacement="")
return(coefs)
}

ylabel <- data.frame(type=c("det","state"), label=c("detection","occupancy"))
plot.marginal <- function(sp, mod, par, par.name, type)
{
  nd.og <- nd.mean[,names(nd.mean) != par] %>% slice(rep(1:n(), times=1000))
  nd <- data.frame(nd.og,
                   par=seq(min(site.stacked[,par]), max(site.stacked[,par]), length.out=1000))
  names(nd) <- gsub(names(nd), pattern="par", replacement=par)
  
  marginal <- predict(mod, type=type, newdata=nd)
  
  nd$StartTime <- times(nd$StartTime) #convert start time back to character
  
  o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
    geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.5, fill="grey75") +
    geom_rug(data=site.stacked, aes(x=site.stacked[,par], y=0)) +
    ylim(0,1) +
    xlab(par.name) + ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +
    geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
  print(o)
  ggsave(o, filename=paste("../Figures/occupancy_BBS/", sp, "_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")
  return(list(plot=o, data=nd, marginal=marginal))
}

plot.marginal2 <- function(sp, mod, par, par.name, type)
{
  nd.og <- nd.mean[,names(nd.mean) != par] %>% slice(rep(1:n(), times=1000))
  nd <- data.frame(nd.og,
                   par=seq(min(site.stacked[,par]), max(site.stacked[,par]), length.out=1000))
  names(nd) <- gsub(names(nd), pattern="par", replacement=par)
  nd[,paste(par, 2, sep="")] <- scale(nd[,par])^2
  
  marginal <- predict(mod, type=type, newdata=nd)
  
  nd$StartTime <- times(nd$StartTime) #convert start time back to character
  
  o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
    geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.5, fill="grey75") +
    geom_rug(data=site.stacked, aes(x=site.stacked[,par], y=0)) +
    ylim(0,1) +
    xlab(par.name) + ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +
    geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
  print(o)
  ggsave(o, filename=paste("../Figures/occupancy_BBS/", sp, "_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")
  return(list(plot=o, data=nd, marginal=marginal))
}

```

## load data
Created detection history data files for each species in 2a_BBS_nightjar_detection_timing.Rmd

```{r}
sp <- "ewpw"

data <- read.table(paste("../Data_processed/BBS_2022/BBS_", sp, ".txt", sep=""), sep=",", header=T)
summary(data)

table(data$Year)
table(data$Month)
table(data[,c("Year","StopTotal")])

#fix start time
data[data$StartTime == 0,]$StartTime <- NA
data$StartTime <- times(paste(substr(data$StartTime, 0,1), ":", substr(data$StartTime, 2,3), ":00", sep=""))

data$StartTemp.F <- as.numeric(data$StartTemp.F)
data$EndTemp.F <- as.numeric(data$EndTemp.F)

#remove rows with incomplete data
data <- data[!is.na(data$StartTemp.F),]
# data <- data[!is.na(data$EndTemp.F),] #not using this

data <- data[!is.na(data$StartTime),]


#remove years before 1997 since hardly any detections then
data <- data[data$Year >= 1997,]

#####
#correct class
data$StateAbb <- factor(data$StateAbb)
data$Month <- factor(data$Month)

data$BCR <- factor(data$BCR)

```

## create umf for stacked single-season occupancy model
```{r}
sp.stacked <- data[,grep(names(data), pattern="Stop[0-9]")]
sp.stacked[sp.stacked > 1 & !is.na(sp.stacked)] <- 1 #convert > 1 to 1 (presence/absence instead of counts)

#check detection histories
unique(sp.stacked)

colSums(sp.stacked) #can we cut off detections anywhere? aka reduce the number of stops
ggplot(data=data.frame(colSums(sp.stacked)), aes(x=1:50, y=colSums.sp.stacked.)) + geom_point(size=3) + 
  geom_hline(aes(yintercept=0), col="grey75", lty="dashed") +
  ylab("Num. detections") + xlab("stop") + theme_bw(base_size=20) + theme(panel.grid = element_blank())
ggsave(filename = "../Figures/occupancy_BBS/EWPW_detectionsbystop.tiff", height=3, width=9, units="in", dpi=400, compression="lzw")

sp.stacked <- sp.stacked[,1:10] #only take the first 10 stops since others are surveys mid-day

######
#siteCovs
site.stacked <- data[,!names(data) %in% grep(names(data), pattern="Stop", value=T)] #siteCovs

#add in transformations for variables, if any
#
#
site.stacked$StartTime2 <- scale(site.stacked$StartTime)^2
site.stacked$jday2 <- scale(site.stacked$jday)^2

site.stacked$Latitude2 <- scale(site.stacked$Latitude)^2
site.stacked$Longitude2 <- scale(site.stacked$Longitude)^2

######
#obsCovs
stop.df <- matrix(rep(1:length(names(sp.stacked)), times=nrow(sp.stacked)), byrow=T, ncol=length(names(sp.stacked))) #add obsCovs, dimensions the same as y data

umf <- unmarkedFrameOccu(y=sp.stacked, siteCovs=site.stacked, obsCovs=list(stop=stop.df))
head(umf)

# tiff(filename=paste("../Figures/occupancy_BBS/",sp,"dethist_BBS_stacked.tiff", sep=""), height=10, width=8, units="in", compression="lzw", res=400)
# plot(umf)
# dev.off()

plot(umf)
summary(umf)

1660/(1660 + 161710) #detected during 1.02% of the surveys
```


## run detection models for stacked single season occupancy models
16351 sites
Maximum number of observations per site: 10 
Mean number of observations per site: 10 
Sites with at least one detection: 1174 

Tabulation of y observations:
     0      1 
161850   1660 

```{r}
Sys.time()
null <- occu(~1 ~1, data=umf)
null.r <- occu(~(1|ct.st.rt) ~1, data=umf) #+ (1|StateAbb)

#obsCov
p.stop <- occu(~stop + (1|ct.st.rt) ~1, data=umf)

#siteCov
p.stemp <- occu(~scale(StartTemp.F) + (1|ct.st.rt) ~1, data=umf)
# p.etemp <- occu(~EndTemp.F + (1|ct.st.rt) ~1, data=umf)

p.month <- occu(~Month + (1|ct.st.rt) ~1, data=umf)
p.jday <- occu(~scale(jday) + (1|ct.st.rt) ~1, data=umf)
p.jday2 <- occu(~scale(jday) + jday2 + (1|ct.st.rt) ~1, data=umf)

p.wind <- occu(~scale(StartWind) + (1|ct.st.rt) ~1, data=umf)
p.sky <- occu(~scale(StartSky) + (1|ct.st.rt) ~1, data=umf)

p.time <- occu(~scale(StartTime) + (1|ct.st.rt) ~1, data=umf)
p.time2 <- occu(~scale(StartTime) + I(scale(StartTime)*scale(StartTime)) + (1|ct.st.rt) ~1, data=umf)
Sys.time()

#model
p.fl <- fitList(null, null.r,
                p.stop, #per minute within a survey
                p.stemp, #p.etemp, 
                p.month, p.jday, p.jday2,
                p.wind, p.sky, #site covariates
                p.time, p.time2) #timing
modSel(p.fl)
#         nPars      AIC   delta AICwt cumltvWt
# p.stop      3  9642.95    0.00  1.00     1.00
# p.time2     4 13695.91 4052.96  0.00     1.00
# p.jday2     4 13724.72 4081.78  0.00     1.00
# p.time      3 13730.92 4087.97  0.00     1.00
# p.stemp     3 13734.61 4091.66  0.00     1.00
# p.month     4 13744.96 4102.01  0.00     1.00

# null.r      2 13751.74 4108.79  0.00     1.00
# p.sky       3 13753.30 4110.35  0.00     1.00
# p.wind      3 13753.52 4110.57  0.00     1.00
# null        2 17176.25 7533.30  0.00     1.00
# p.jday      3 18517.78 8874.83  0.00     1.00

Sys.time()
p.global <- occu(~scale(stop) + scale(StartTime) + StartTime2 +
                   scale(jday) + jday2 +
                   scale(StartTemp.F) + # scale(StartWind) +
                   (1|ct.st.rt) + (1|StateAbb) ~1, data=umf)

p.global1 <- occu(~scale(stop) + scale(StartTime) + I(scale(StartTime)^2) +
                   scale(jday) + jday2 +
                   scale(StartTemp.F) + # scale(StartWind) +
                   (1|ct.st.rt) + (1|StateAbb) ~1, data=umf)
p.global2 <- occu(~scale(stop) + scale(StartTime) + StartTime2 +
                   scale(jday) + I(scale(jday)^2) +
                   scale(StartTemp.F) + # scale(StartWind) +
                   (1|ct.st.rt) + (1|StateAbb) ~1, data=umf)


```

## coefficients and marginal plots for detection variables
```{r}
#these look kind of weird because need to set random effect variables
plotEffects(p.stop, "det", "stop")
plotEffects(p.stemp, "det", "StartTemp.F")
plotEffects(p.etemp, "det", "EndTemp.F")
plotEffects(p.wind, "det", "StartWind")
plotEffects(p.sky, "det", "StartSky")
plotEffects(p.month, "det", "Month")
plotEffects(p.time, "det", "StartTime")
plotEffects(p.time2, "det", "StartTime")

p.coefs <- vals.coefs(p.global) #choose model to extract coefficients from

c.plot <- ggplot(data=p.coefs[!p.coefs$name %in% "Int",], aes(x=name, y=est, group=p, shape=p)) +
  # c.plot <- ggplot(data=coefs[coefs$name != "INT",], aes(x=name, y=est)) +
  geom_point(size=3, position=position_dodge(width=0.5)) + 
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.5)) +
  # geom_errorbar(aes(ymin=est-se, ymax=est+se), width=0, lwd=1, position=position_dodge(width=0.5)) +
  geom_hline(aes(yintercept=0), lty="dotted") +
  xlab("parameter") + ylab("estimate") +
  coord_flip() +
  theme_bw(base_size=16) + theme(panel.grid=element_blank()) + facet_wrap(.~p, scales="free", ncol=1)
c.plot

ggsave(c.plot, filename="../Figures/occupancy_BBS/EWPW_stacked_occupancy_det_coefficients.tiff", height=6, width=6, units="in", dpi=400, compression="lzw")

# ###########
# #manually plot marginals in ggplot2
nd.mean <- site.stacked %>% mutate_if(is.numeric, mean)
nd.mean <- nd.mean[100,] #only take one row since all rows are the same (which row matters, since random effect by route and state)
nd.mean$stop <- 1

#
#plot these with p.global2, need to pull out scale() from within I() to plot these
plot.marginal(sp="EWPW", p.global, "StartTemp.F", "start temperature (F)", type="det")
# plot.marginal(sp="EWPW", p.global, "StartWind", "wind", type="det")
plot.marginal(sp="EWPW", p.global2, "jday", "Julian day", type="det")

#Start time is a little weird since using chron package
Sys.setenv(TZ='GMT')

type <- "det"
st.time <- plot.marginal(sp="EWPW", p.global1, "StartTime", "Start Time", type="det")
  st.time + scale_x_chron(format="%H:%M", n=10)
  ggsave(filename=paste("../Figures/occupancy_BBS/", toupper(sp), "_stacked_occupancy_",type,"_", "StartTime",".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")
  

#for stop
mod <- p.global; par <- "stop"; par.name <- "stop number"; type <- "det"
nd.og <- nd.mean[,names(nd.mean) != par] %>% slice(rep(1:n(), times=10))
nd <- data.frame(nd.og, par=1:10)
names(nd) <- gsub(names(nd), pattern="par", replacement=par)
marginal <- unmarked::predict(mod, type=type, newdata=nd)
o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.5, fill="grey75") +
  ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +ylim(0,1) +
  xlab(par.name) + scale_x_continuous(breaks=1:10) +
  geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
print(o)
ggsave(o, filename=paste("../Figures/occupancy_BBS/EWPW_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")


```


## run occupancy models for stacked single season occupancy models
```{r}

null <- occu(~1 ~1, data=umf)
null.r <- occu(~1 ~(1|ct.st.rt), data=umf)

o.state <- occu(~1 ~ StateAbb + (1|ct.st.rt), data=umf)
o.bcr <- occu(~1 ~ BCR + (1|ct.st.rt), data=umf)
o.lat <- occu(~1 ~ scale(Latitude) + (1|ct.st.rt), data=umf)
o.lat2 <- occu(~1 ~ scale(Latitude) + I(scale(Latitude)^2) + (1|ct.st.rt), data=umf)
o.long <- occu(~1 ~ scale(Longitude) + (1|ct.st.rt), data=umf)
o.long2 <- occu(~1 ~ scale(Longitude) + I(scale(Longitude)^2) + (1|ct.st.rt), data=umf)

o.year <- occu(~1 ~ scale(Year) + (1|ct.st.rt), data=umf)
o.year.f <- occu(~1 ~ factor(Year) + (1|ct.st.rt), data=umf)

o.fl <- fitList(null, null.r,
                o.state, o.bcr, o.lat, o.lat2, o.long, o.long2,
                o.year, o.year.f)
modSel(o.fl)

#          nPars      AIC   delta   AICwt cumltvWt
# o.state     18 13836.93    0.00 9.8e-01     0.98 #crazy SEs for VT and FL
# o.lat2       4 13844.95    8.02 1.8e-02     1.00
# o.year       3 13848.97   12.03 2.4e-03     1.00
# o.bcr        8 13873.61   36.68 1.1e-08     1.00
# o.long2      4 13885.61   48.67 2.6e-11     1.00

# null.r       2 13888.35   51.42 6.7e-12     1.00
# o.lat        3 13889.23   52.29 4.3e-12     1.00
# o.year.f    25 14226.55  389.62 2.4e-85     1.00
# null         2 17176.25 3339.32 0.0e+00     1.00
# o.long       3 18063.08 4226.15 0.0e+00     1.00

o.year.state <- occu(~1 ~ scale(Year) + StateAbb + (1|ct.st.rt), data=umf)
o.year.bcr <- occu(~1 ~ scale(Year) + BCR + (1|ct.st.rt), data=umf)
# o.year.state.bcr <- occu(~1 ~ scale(Year) + StateAbb + BCR + (1|ct.st.rt), data=umf) #this won't run

o.global0 <- occu(~1 ~ scale(Latitude) + Latitude2 + 
                    scale(Longitude) + Longitude2 + 
                    scale(Year) + # BCR + #can't add in BCR, will not converge
                    (1|ct.st.rt), data=umf)
o.global1 <- occu(~1 ~ scale(Latitude) + I(scale(Latitude)^2) + 
                    scale(Longitude) + Longitude2 + 
                    scale(Year) +
                    (1|ct.st.rt), data=umf)
o.global2 <- occu(~1 ~ scale(Latitude) + Latitude2 + 
                    scale(Longitude) + I(scale(Longitude)^2) + 
                    scale(Year) +
                    (1|ct.st.rt), data=umf)


o.fl <- fitList(null, null.r,
                o.year.state, o.year.bcr, o.global1, # o.year.state.bcr, 
                o.state, o.bcr, o.lat, o.lat2, o.long, o.long2,
                o.year, o.year.f)
modSel(o.fl)

#              nPars      AIC   delta   AICwt cumltvWt
# o.year.state    19 13795.08    0.00 9.8e-01     0.98 #this is weird
# o.global1        7 13803.04    7.96 1.8e-02     1.00
# o.year.bcr       9 13832.42   37.34 7.7e-09     1.00
# o.state         18 13836.93   41.85 8.0e-10     1.00
# o.lat2           4 13844.95   49.87 1.5e-11     1.00
# o.year           3 13848.96   53.88 2.0e-12     1.00
# o.bcr            8 13873.61   78.53 8.7e-18     1.00
# o.long2          4 13885.61   90.53 2.2e-20     1.00
# null.r           2 13888.35   93.27 5.5e-21     1.00
# o.lat            3 13889.23   94.15 3.5e-21     1.00
# o.year.f        25 14226.55  431.47 2.0e-94     1.00
# null             2 17176.25 3381.17 0.0e+00     1.00
# o.long           3 18063.08 4268.00 0.0e+00     1.00

```

## coefficients and marginal plots for occupancy variables
```{r}
nd.mean <- site.stacked %>% mutate_if(is.numeric, mean)
nd.mean <- nd.mean[100,] #only take one row since all rows are the same (which row matters, since random effect by route and state)

#continuous variables
# plot.marginal(sp="EWPW", o.lat2, "Latitude", "Latitude", type="state")
# plot.marginal(sp="EWPW", o.long2, "Longitude", "Longitude", type="state")

# plot.marginal(sp="EWPW", o.year.state, "Year", "Year", type="state")
y <- plot.marginal(sp="EWPW", o.global0, "Year", "Year", type="state")
range(y$marginal$Predicted)

plot.marginal(sp="EWPW", o.global1, "Latitude", "Latitude", type="state")
plot.marginal(sp="EWPW", o.global2, "Longitude", "Longitude", type="state")

#categorical variables
plotEffects(o.state, "StateAbb", type="state")
plotEffects(o.bcr, "BCR", type="state")

######
#plot values for each state
######
mod <- o.year.state
nd <- unique(site.stacked[,c("StateAbb","BCR","Year","ct.st.rt")])
nd <- nd %>% #take first row of each group
  group_by(StateAbb, Year) %>%
  arrange(ct.st.rt) %>%
  filter(row_number()==1)
marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$Year, m.data$StateAbb),]
m.data$StateAbb <- factor(m.data$StateAbb, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","RI","MA","VT","NH","ME")))

# ggplot(m.data, aes(x=StateAbb, y=Predicted)) +
ggplot(m.data, aes(x=Year, y=Predicted)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, col="black", position=position_dodge(width=0.25)) +
  xlab("state") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  facet_wrap(~StateAbb, ncol=2, dir="v") +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy_BBS/EWPW_stacked_occupancy_psi_state-year.tiff", height=20, width=8, units="in", dpi=400, compression="lzw")
```

## run det + occ models
```{r}
p.global <- occu(~scale(stop) + scale(StartTime) + StartTime2 +
                   scale(jday) + jday2 +
                   scale(StartTemp.F) + # scale(StartWind) +
                   (1|ct.st.rt) ~1, data=umf)
o.global <- occu(~1 ~ scale(Latitude) + Latitude2 + 
                    scale(Longitude) + Longitude2 + 
                    scale(Year) + # BCR + #can't add in BCR, will not converge
                    (1|ct.st.rt), data=umf)

po.global <- occu(~ scale(stop) + scale(StartTime) + StartTime2 +
                   scale(jday) + jday2 +
                   scale(StartTemp.F) + # scale(StartWind) +
                   (1|ct.st.rt) 
                  ~ scale(Latitude) + Latitude2 + 
                    scale(Longitude) + Longitude2 + 
                    scale(Year) + # BCR + #can't add in BCR, will not converge
                    (1|ct.st.rt), data=umf)
modSel(fitList(p.global, o.global0, po.global))

#           nPars      AIC   delta   AICwt cumltvWt
# po.global    13  9332.82    0.00 1.0e+00     1.00
# p.global      8  9581.30  248.47 1.1e-54     1.00
# o.global0     7 13803.04 4470.22 0.0e+00     1.00

summary(po.global)

# Call:
# occu(formula = ~scale(stop) + scale(StartTime) + StartTime2 + 
#     scale(jday) + jday2 + scale(StartTemp.F) + (1 | ct.st.rt) ~ 
#     scale(Latitude) + Latitude2 + scale(Longitude) + Longitude2 + 
#         scale(Year) + (1 | ct.st.rt), data = umf)
# 
# Occupancy (logit-scale):
# Random effects:
#    Groups        Name Variance Std.Dev.
#  ct.st.rt (Intercept)   185.21   13.609
# 
# Fixed effects:
#                  Estimate     SE       z   P(>|z|)
# (Intercept)       -10.059 0.5234 -19.218  2.62e-82
# scale(Latitude)    -1.641 0.7168  -2.289  2.21e-02
# Latitude2          -1.890 0.6460  -2.925  3.44e-03
# scale(Longitude)    0.270 0.5062   0.534  5.93e-01
# Longitude2          0.329 0.3474   0.947  3.43e-01
# scale(Year)        -1.827 0.0734 -24.878 1.28e-136
# 
# Detection (logit-scale):
# Random effects:
#    Groups        Name Variance Std.Dev.
#  ct.st.rt (Intercept)    1.471    1.213
# 
# Fixed effects:
#                    Estimate     SE      z  P(>|z|)
# (Intercept)         -5.6221 0.1403 -40.07 0.00e+00
# scale(stop)         -3.0101 0.0775 -38.85 0.00e+00
# scale(StartTime)    -0.2351 0.1392  -1.69 9.12e-02
# StartTime2          -0.3013 0.0999  -3.02 2.57e-03
# scale(jday)         -0.2741 0.0391  -7.01 2.43e-12
# jday2                0.1557 0.0218   7.15 8.47e-13
# scale(StartTemp.F)  -0.0731 0.0382  -1.91 5.60e-02
# 
# AIC: 9332.824 
# Number of sites: 16337
# optim convergence code: 0
# optim iterations: 149 
# Bootstrap iterations: 0 

plot.marginal2(sp="global/EWPW", po.global, "Longitude", "Longitude", type="state")

#det
plot.marginal(sp="global/EWPW", po.global, "StartTemp.F", "Temperature (F)", type="det")
plot.marginal2(sp="global/EWPW", po.global, "jday", "Julian day", type="det")

st.time <- plot.marginal2(sp="EWPW", po.global, "StartTime", "Start Time", type="det")
  st.time + scale_x_chron(format="%H:%M", n=10)
  ggsave(filename=paste("../Figures/occupancy_BBS/global/", toupper(sp), "_stacked_occupancy_",type,"_", "StartTime",".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")



```
