==> MarkdownDocs/4_survey_locations_for_nightjars_2023-06-29.Rmd <==
---
title: "4_survey locations for nightjars"
author: "Marie I. Tosa"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#load packages
require(plyr)
require(tidyr)
require(units)

#for GIS
require(sf)
require(raster)
require(stars)
require(terrain)

#for plotting
require(ggplot2)
require(ggrepel)

AF_states <- read_sf(dsn="../../GIS", layer="AF_states")

BCR.ST <- read_sf(dsn="../../GIS", layer="BCR-state") #area in mi^2
unique(BCR.ST$ST)
BCR.ST$ST <- factor(BCR.ST$ST)
BCR.ST$BCR <- factor(BCR.ST$BCR)
```

## Load road data

Have to go county by county with TIGER/LINE roads

Download road layers by state from state websites:
ME - https://www.maine.gov/geolib/catalog.html#transportation
MA - https://www.mass.gov/info-details/massgis-data-massachusetts-department-of-transportation-massdot-roads

```{r, echo=T}
# road.files <- dir("../../GIS/tigerline_roads/", pattern=".shp$", full.names=F)
# 
# for(f in roadfiles)
# {
#   print(f)
#   r <- read_sf(dsn="../../GIS/tigerline_roads", layer=gsub(f, pattern=".shp", replacement=""))
# }

#try downloading at state scale
# me.rd <- read_sf(dsn="../../GIS/roads/MaineDOT_Public_Roads/", layer="MaineDOT_Public_Roads")
# ma.rd <- read_sf(dsn="../../GIS/roads/MassDOT_Roads_SHP/", layer="EOTROADS_ARC")
#no CLASS=0,1,2


# ggplot(data=ma.rd[ma.rd$CLASS %in% 4:6 & ma.rd$LENGTH_MI > 1,]) + geom_sf(aes(col=CLASS)) + theme_bw() + theme(panel.grid=element_blank())

```

## Load national land cover data
```{r, echo=T}
# nlcd_2019 <- raster("../../GIS/NLCD/nlcd_2019_land_cover_l48_20210604/nlcd_2019_land_cover_l48_20210604.img")
# AF_states <- readOGR(dsn="../../GIS", layer="AF_states")
# 
# nlcd_2019 <- crop(nlcd_2019, spTransform(AF_states, CRSobj = projection(nlcd_2019)))
# nlcd_2019 <- mask(nlcd_2019, spTransform(AF_states, CRSobj = projection(nlcd_2019)))
# writeRaster(nlcd_2019, filename = "../../GIS/NLCD/AF_nlcd_2019.tif", format="GTiff", overwrite=T)

nlcd_2019 <- raster("../../GIS/NLCD/AF_nlcd_2019.tif")

```

## Load national land cover data
```{r, echo=T}
# tcc <- raster("../../GIS/nlcd_tcc_CONUS_all/nlcd_tcc_CONUS_2021_v2021-4/nlcd_tcc_conus_2021_v2021-4.tif")
# AF_states <- readOGR(dsn="../../GIS", layer="AF_states")
# tcc_2019 <- crop(tcc, spTransform(AF_states, CRSobj = projection(tcc)))
# tcc_2019 <- mask(tcc_2019, spTransform(AF_states, CRSobj = projection(tcc)))
# writeRaster(tcc_2019, filename = "../../GIS/nlcd_tcc_CONUS_all/AF_tcc_2019.tif", format="GTiff", overwrite=T)

tcc_2019 <- raster("../../GIS/nlcd_tcc_CONUS_all/AF_tcc_2019.tif")
# tcc_2019.df <- as.data.frame(tcc_2019, xy=T) #too big, can't do
# table(tcc_2019.df$)

```

## Randomly distribute points in each BCR
only do this once, since it will generate new points each time
```{r, echo=T}

comp <- read.csv(file="../Data_processed/sampling_comparison_nsn_bbs.csv")
comp <- comp[!is.na(comp$BCR),]

#can you set so points have to be a certain distance away from each other? Yes.
buffer <- 20*1609 #20 miles in m
multiplier <- 2
route.pts <- NULL

for(r in 1:nrow(comp))
{
  a <- comp[r,]
  print(paste(a$ST, a$BCR))
  if(a$nsn3k.area.routes == 0)
  {
    next
  }
  a.sf <- BCR.ST[BCR.ST$ST == a$ST & BCR.ST$BCR == a$BCR,]
  # a.pts <- st_sample(a.sf, size=a$nsn3k.area.routes) #sample without buffer around each point
  
  a.laea <- st_transform(a.sf, #convert projection to run through SSI
      crs = "+proj=laea +x_0=4600000 +y_0=4600000 +lon_0=0.13 +lat_0=0.24 +datum=WGS84 +units=m")
  # a.pts <- st_sample(x=st_as_sfc(st_bbox(a.laea)), type="SSI", r=1000, n=100) #this works
  a.pts <- st_sample(x=a.laea, type="SSI", r=buffer, n=a$nsn3k.area.routes*multiplier) #needs crs
  
  a.pts <- data.frame(st_coordinates(a.pts))
  a.pts <- a.pts[!duplicated(a.pts),]
  
  a.pts.sf <- st_as_sf(data.frame(a.pts), coords=c("X","Y"), crs=crs(a.laea))
  a.pts.latlong <- st_transform(a.pts.sf, crs=crs(a.sf))
  a.pts.coords <- st_coordinates(a.pts.latlong)
  
  a.pts <- data.frame(a.pts.coords, ST=a$ST, BCR=a$BCR)
  
  # a.pts <- st_set_crs(a.pts, crs(a.laea))
  # a.pts.latlong <- st_transform(a.pts, crs(a.sf))
  # 
  # a.pts <- st_coordinates(a.pts.latlong)
  # a.pts <- data.frame(a.pts, ST=a$ST, BCR=a$BCR)
  
  a.pts$selected <- 0
  rows <- sample(1:nrow(a.pts), size=a$nsn3k.area.routes, replace=F) #if do this within [,] something weird happens and duplicates rows
  a.pts[rows,]$selected <- 1
  
  route.pts <- rbind(route.pts, a.pts)
}

route.pts$selected <- factor(route.pts$selected)

nrow(route.pts[duplicated(route.pts[,c("X","Y")]),]) #check to see if any points are duplicated

route.sf <- st_as_sf(route.pts, coords=c("X","Y"), crs=crs(BCR.ST))
BCR.ST$BCR <- factor(BCR.ST$BCR)
for(s in unique(AF_states$ST))
{
  ex.s <- extent(AF_states[AF_states$ST == s,])
  ggplot() + geom_sf(data=BCR.ST, aes(fill=BCR)) + 
        scale_fill_brewer(palette="Set2") +
    geom_sf(data=route.sf[route.sf$ST != s,], size=1, aes(shape=selected, col=selected)) +
    geom_sf(data=route.sf[route.sf$ST == s,], size=3, aes(shape=selected, col=selected)) +
    scale_shape_manual(values=c(1,19)) + scale_color_manual(values=c("gray50","black")) +
    coord_sf(xlim=c(ex.s@xmin, ex.s@xmax), ylim=c(ex.s@ymin, ex.s@ymax)) +
    theme_bw(base_size=20) + theme(panel.grid=element_blank())
  # ggsave(filename = paste("../Figures/random_points_3k/randompts_", s, ".tiff", sep=""),
  #        height=8, width=8, units="in", dpi=400, compression="lzw")
}
# write.table(route.pts, file="../Data_processed/random_pts_area3k.txt", sep=",", row.names=F)
#EPSG:4269
```
## (Start Here) Determining whether we can keep existing routes
compare with current survey locations
```{r, echo=T}
random.routes <- read.table("../Data_processed/random_pts_area3k.txt", sep=",", header=T)
random.routes$route_name <- paste(random.routes$BCR, random.routes$ST, 1:nrow(random.routes), sep="-")
routes <- read.table("../Data_processed/master_survey_locations.txt", sep=",", header=T) #all nightjar survey routes
routes$st.route.stop <- paste(routes$st.route, routes$Stop, sep="..")

random.sf <- st_as_sf(random.routes, coords=c("X","Y"), crs=crs(BCR.ST))
routes.sf <- st_as_sf(routes[!is.na(routes$Latitude),], coords=c("Longitude","Latitude"), crs=crs(BCR.ST))

# ggplot(data=routes.sf[routes.sf$state=="ME",]) + geom_sf()

#calculate distance between points
dist.routes <- st_distance(random.sf, routes.sf, by_element=F)
units(dist.routes) <- make_units(mi) #convert units from m to mi
# units(dist.routes) <- as_units(quote(mi)) #convert units from m to mi
#rows = random.sf, columns=routes.sf

#convert from wide to long
dist.routes <- data.frame(dist.routes)
names(dist.routes) <- routes[!is.na(routes$Latitude),]$st.route.stop #label columns
dist.routes <- cbind(dist.routes, random.routes[,c("route_name","selected")]) #label rows

head(names(dist.routes))
tail(names(dist.routes)) #make sure the next line of code contains the first and last route names
dist.long <- pivot_longer(dist.routes, values_to="dist", cols='FL_0001..1':'MA_Quinapoxet River, Holden (Forestry lot scouting observations)..6') 
dist.long <- separate(dist.long, col="name", into=c("st_route","stop"), remove=F, sep="[.][.]")

### change this as necessary ###
distance.threshold <- set_units(20, mi)
###

to_combine <- dist.long[dist.long$dist < distance.threshold,]
to_combine <- to_combine[order(to_combine$stop),]
to_combine <- to_combine[!duplicated(to_combine[,c("route_name","st_route")]),]

#add lat long coordinates to route_name and st_route_stop
to_combine <- merge(to_combine, random.routes, by=c("route_name","selected"), all.x=T)
to_combine <- merge(to_combine, 
                    routes[,c("st.route.stop","Longitude","Latitude")],
                    by.x="name", by.y="st.route.stop", all.x=T, suffixes=c(".random",".current"))
# name = st.route.stop in routes
# route_name = route_name in random.routes
# st_route = st.route in routes
# X and Y = coordinates of random.routes
# Longitude and Latitude = coordinates of current NSN routes

to_combine <- to_combine[to_combine$ST == substr(to_combine$st_route, 0, 2),] #make sure existing and random routes are in the same state

###########
#how many random.routes already have existing routes that could be used
sum.r <- ddply(to_combine, .(route_name, selected, ST, BCR),
               summarize, num.existing=length(name)) #summarized by random.route
sum.bcr.st <- ddply(sum.r, .(selected, ST, BCR), summarize, num.random.routes=length(route_name)) #summarize by bcr-state level, number of random routes with existing routes

#need to compare with how many routes we need run
need.r <- ddply(random.routes, .(ST, BCR, selected), summarize, num.random.need=length(route_name))

summary.need.have <- merge(need.r, sum.bcr.st, by=c("ST","BCR","selected"), all.x=T)
summary.need.have[is.na(summary.need.have)] <- 0
summary.need.have$num.remaining <- summary.need.have$num.random.need - summary.need.have$num.random.routes

write.csv(summary.need.have, file="../Data_processed/random_pts_area3k_vs_existingroutes.csv", row.names=F) #isn't quite right because has points in different BCRs

# to_combine$selected <- factor(to_combine$selected)
# random.sf$selected <- factor(random.sf$selected)

#can't seem to get the symbols and shapes correct with "selected" column
# to_combine$alternate <- 1-to_combine$selected
# random.sf$alternate <- 1-random.sf$selected
# 
# to_combine$alternate <- factor(to_combine$alternate)
# random.sf$alternate <- factor(random.sf$alternate)

#############
#create tables for each state with existing routes and random routes that need to be created
#############
names(to_combine) <- c("existing_rt_stop","random_route_name","selected","existing_route","stop","distance",
                       "X.random","Y.random","state.random","BCR.random","Longitude.existing","Latitude.existing") #,"alternate")
# to_combine <- separate(to_combine, col="existing_route", into=c("state.existing","existing_route_name"), sep="_", remove=F)
# to_combine <- to_combine[to_combine$state.random == to_combine$state.existing,] #make sure existing and random routes are in the same state

#if existing route listed multiple times, remove the one that is further away
to_combine <- to_combine[order(to_combine$existing_route, to_combine$distance),]
to_combine <- to_combine[!duplicated(to_combine$existing_route),]

state_table <- merge(to_combine[,c("existing_rt_stop","existing_route","stop",
                                   "Longitude.existing","Latitude.existing",
                                   "distance","random_route_name","selected")], 
                     random.routes, by.x=c("random_route_name","selected"), by.y=c("route_name","selected"), all.y=T)
state_table <- state_table[order(state_table$ST, state_table$BCR, state_table$selected),]

names(state_table) <- c("random_route_name","selected","existing_rt_stop","existing_route","stop",
                        "Longitude.existing","Latitude.existing","distance","X.random","Y.random","ST","BCR")

write.csv(state_table[,c("random_route_name","selected","existing_route","Longitude.existing","Latitude.existing","distance","X.random","Y.random","ST","BCR")], file="../Data_processed/table_for_states.csv", row.names=F)

########
s1 <- ddply(state_table, .(random_route_name, ST, BCR, selected), 
            summarize, num.existing=sum(!is.na(existing_route)), num.na=sum(is.na(existing_route)))
s2 <- ddply(s1, .(ST, BCR, selected), summarize, num.exist=sum(num.existing != 0), num.need=sum(num.na), num.total=length(num.existing))

write.csv(s2, file="../Data_processed/table_for_states_summary.csv", row.names=F)

#############
#plot maps
#############
require(ggspatial)

for(st in unique(AF_states$ST))
{
  print(st)
  #create sf objects from to_combine
  # random.st.sf <- st_as_sf(to_combine[to_combine$ST == st,], coords=c("X","Y"), crs=crs(BCR.ST))
  existing.sf <- st_as_sf(to_combine[to_combine$state.random == st,], coords=c("Longitude.existing","Latitude.existing"), crs=4269)
  # all.random.sf <- random.sf[random.sf$ST == st &
  #                              !random.sf$route_name %in% random.st.sf$route_name,]

  
  random.st.sf <- st_as_sf(to_combine, coords=c("X.random","Y.random"), crs=4269)
  # existing.sf <- st_as_sf(to_combine, coords=c("Longitude","Latitude"), crs=crs(BCR))

  all.random.sf <- random.sf[random.sf$ST == st &
                               !random.sf$route_name %in% to_combine[to_combine$state.random == st,]$random_route_name,]
  
  #plot
  ex <- extent(AF_states[AF_states$ST == st,])
  plot <- ggplot() + 
    # geom_sf(data=AF_states[AF_states$ST == st,], fill=NA, col="grey75") + 
    geom_sf(data=BCR.ST, aes(fill=BCR), alpha=0.5) +
    scale_fill_brewer(palette="Set2") +
    
    # geom_sf(data=all.random.sf, aes(col=selected), pch=17, size=3) + #random points
    
    # geom_sf(data=random.st.sf, aes(col=selected, shape=selected), stroke=1.5) +
    # scale_shape_manual(values=c(1,19)) + scale_color_manual(values=c("grey50","black")) +
    scale_shape_manual(values=c(19,1)) + scale_color_manual(values=c("black","grey50")) +
    
    # geom_sf(data=existing.sf, col="steelblue", aes(shape=selected)) + 
    # geom_text_repel(data=existing.sf, aes(label=st_route, geometry=geometry, col=selected), stat = "sf_coordinates") +
    # coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax)) +
    xlab("") + ylab("") +
    annotation_scale(unit_category = "imperial") +
    theme_bw(base_size=20) + theme(panel.grid=element_blank())
  
  #only include points that were selected
  plot + 
    geom_sf(data=all.random.sf[all.random.sf$selected == 1,], aes(col=alternate, shape=alternate), pch=17, size=3) + #triangle for random points without existing nearby
    geom_sf(data=random.st.sf[random.st.sf$selected == 1,], aes(col=alternate, shape=alternate), stroke=1.5) + 
    geom_sf(data=existing.sf[existing.sf$selected == 1,], col="steelblue", aes(shape=alternate)) + 
    geom_text_repel(data=existing.sf[existing.sf$selected == 1,],
                    aes(label=existing_route, geometry=geometry, col=alternate), stat = "sf_coordinates", col="steelblue") +
    coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax))
  ggsave(filename=paste("../Figures/random_3k_with_existing_no_alternates/random_existing_noalternates_",st,".tiff", sep=""),
         height=18, width=18, units="in", dpi=400, compression="lzw")
  
  #all selected and alternate points
  plot + 
    geom_sf(data=all.random.sf, aes(col=alternate), pch=17, size=5) +
    geom_sf(data=random.st.sf, aes(col=alternate, shape=alternate), stroke=1.5) +
    geom_sf(data=existing.sf, col="steelblue", aes(shape=alternate)) + scale_shape_manual(values=c(19,1)) +
    geom_text_repel(data=existing.sf, aes(label=existing_route, geometry=geometry, col=alternate), stat = "sf_coordinates") +
    coord_sf(xlim=c(ex@xmin, ex@xmax), ylim=c(ex@ymin, ex@ymax))
  #blue dots = existing points, black dots = randomly distributed locations
  ggsave(filename=paste("../Figures/random_3k_with_existing/random_existing_",st,".tiff", sep=""),
         height=18, width=18, units="in", dpi=400, compression="lzw")
}
```

# compare locations of BBS routes to random locations
```{r}

random.routes <- read.table("../Data_processed/random_pts_area3k.txt", sep=",", header=T)
random.routes$route_name <- paste(random.routes$BCR, random.routes$ST, 1:nrow(random.routes), sep="-")
random.sf <- st_as_sf(random.routes, coords=c("X","Y"), crs=crs(BCR.ST))


#load BBS route locations
rt <- read.csv("../Data_raw/BBS_2022/routes.csv")
  rt <- rt[rt$BCR %in% BCR$BCR,]
  table(rt[,c("Active","BCR")]) #number of routes in each BCR
  table(rt$StateNum) #number of routes in each state
  
  rt$Active <- factor(rt$Active)
  rt$StateNum <- factor(rt$StateNum)
  rt$BCR <- factor(rt$BCR)
  
  rt$ct.st.rt <- paste(rt$CountryNum, rt$StateNum, rt$Route, sep="-")
  
  st.key <- read.csv("../Data_raw/BBS_state_key.csv")
  rt <- merge(rt, st.key[,c("StateNum","StateAbb")], by="StateNum", all.x=T)

rt.sf <- st_as_sf(rt, coords=c("Longitude","Latitude"), crs=4269)

#calculate distance between points
bbs.dist.routes <- st_distance(random.sf, rt.sf, by_element=F)
units(bbs.dist.routes) <- make_units(mi) #convert units from m to mi
# units(dist.routes) <- as_units(quote(mi)) #convert units from m to mi
#rows = random.sf, columns=rt.sf

#convert from wide to long
dist.routes <- data.frame(bbs.dist.routes)
names(dist.routes) <- rt[!is.na(rt$Latitude),]$ct.st.rt #label columns
dist.routes <- cbind(dist.routes, random.routes[,c("route_name","selected")]) #label rows

head(names(dist.routes))
tail(names(dist.routes)) #make sure the next line of code contains the first and last route names
dist.long <- pivot_longer(dist.routes, values_to="dist", cols='840-2-1':'840-90-951') 

### change this as necessary ###
distance.threshold <- set_units(20, mi)
###

to_combine <- dist.long[dist.long$dist < distance.threshold,]

#add lat long coordinates to route_name and st_route_stop
to_combine <- merge(to_combine, random.routes, by=c("route_name","selected"), all.x=T)
to_combine <- merge(to_combine, 
                    rt[,c("ct.st.rt","BCR","StateAbb","Longitude","Latitude")],
                    by.x="name", by.y="ct.st.rt", all.x=T, suffixes=c(".random",".bbs"))
# name = ct.st.rt in rt
# route_name = route_name in random.routes
# st_route = st.route in routes
# X and Y = coordinates of random.routes
# Longitude and Latitude = coordinates of BBS routes

to_combine <- to_combine[to_combine$ST == to_combine$StateAbb,] #make sure existing and random routes are in the same state
to_combine <- to_combine[to_combine$BCR.random == to_combine$BCR.bbs,] #make sure existing and random routes are in the same BCR
names(to_combine) <- gsub(names(to_combine), pattern="BCR.random", replacement="BCR")

###########
#how many random.routes already have existing BBS routes that could be used
sum.r <- ddply(to_combine, .(route_name, selected, ST, BCR),
               summarize, num.existing=length(name)) #summarized by random.route
sum.bcr.st <- ddply(sum.r, .(selected, ST, BCR), summarize, num.random.routes=length(route_name)) #summarize by bcr-state level, number of random routes with existing BBS routes

#need to compare with how many routes we need run
need.r <- ddply(random.routes, .(ST, BCR, selected), summarize, num.random.need=length(route_name))

summary.need.have <- merge(need.r, sum.bcr.st, by=c("ST","BCR","selected"), all.x=T)
summary.need.have[is.na(summary.need.have)] <- 0
summary.need.have$num.remaining <- summary.need.have$num.random.need - summary.need.have$num.random.routes

write.csv(summary.need.have, file="../Data_processed/random_pts_area3k_vs_BBSroutes.csv", row.names=F) #this isn't quite right since includes routes from a different state

#############
#create tables for each state with existing routes and random routes that need to be created
#############
to_combine <- to_combine[,!names(to_combine) %in% c("BCR.bbs","StateAbb")]
names(to_combine) <- c("BBS_route","random_route_name","selected","distance",
                       "X.random","Y.random","state.random","BCR.random","Longitude.BBS","Latitude.BBS")

#if existing route listed multiple times, remove the one that is further away
to_combine <- to_combine[order(to_combine$BBS_route, to_combine$distance),]
to_combine <- to_combine[!duplicated(to_combine[,c("BBS_route","random_route_name")]),]

state_table <- merge(to_combine[,c("BBS_route",
                                   "Longitude.BBS","Latitude.BBS",
                                   "distance","random_route_name","selected")], 
                     random.routes, by.x=c("random_route_name","selected"), by.y=c("route_name","selected"), all.y=T)
state_table <- state_table[order(state_table$random_route_name),]
state_table <- state_table[order(state_table$ST, state_table$BCR, state_table$selected),]

names(state_table) <- c("random_route_name","selected","BBS_route",
                        "Longitude.BBS","Latitude.BBS","distance","X.random","Y.random","ST","BCR")

write.csv(state_table, file="../Data_processed/table_for_states_BBS.csv", row.names=F)

########
s1 <- ddply(state_table, .(random_route_name, ST, BCR, selected), #number of BBS routes per random route
            summarize, num.BBS=sum(!is.na(BBS_route)), num.na=sum(is.na(BBS_route)))
s2 <- ddply(s1, .(ST, BCR, selected), summarize, num.BBS=sum(num.BBS != 0), num.need=sum(num.na), num.total=length(random_route_name))
s2[s2$selected == 1,]$selected <- "selected"
s2[s2$selected == 0,]$selected <- "alternate"

write.csv(s2, file="../Data_processed/table_for_states_summary_BBS.csv", row.names=F)

```
