==> MarkdownDocs/6_nsn_occupancy_unmarked_stacked_EWPW_2023-07-25.Rmd <==
---
title: "6_nightjar surveys_single-season occupancy_stacked years_EWPW"
author: "Marie I. Tosa"
date: "`r Sys.Date()`"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

require(unmarked)
require(plyr)
require(dplyr)
require(tidyr)

require(ggplot2)
require(ggpubr)
```

## Create functions to plot figures
```{r}
vals.coefs <- function(mod)
{
coefs <- coef(mod) %>% tibble::enframe() #extract coefficients for covariates
se <- SE(mod) %>% tibble::enframe() #extract SEs for covariates

ci.psi <- confint(mod, type="state") %>% tibble::enframe()
ci.det <- confint(mod, type="det") %>% tibble::enframe()
ci <- rbind(ci.psi, ci.det) #combine psi and det CIs
ci <- data.frame(name=ci$name, lower=ci$value[,1], upper=ci$value[,2]) #need to split up ci so can read as data frame with lower and upper bounds

coefs <- merge(coefs, se, by="name") #merge coefficient values and ses
coefs <- merge(coefs, ci) #merge coefs with CIs
coefs <- data.frame(coefs)
names(coefs) <- c("par", "est", "se", "lower","upper")

coefs$p <- gsub(substr(coefs$par, 0,3), pattern="p\\([A-z]", replacement="det")
coefs$name <- gsub(coefs$par, pattern="scale\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="p\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="psi\\(", replacement="")
coefs$name <- gsub(coefs$name, pattern="\\)\\)", replacement="")
coefs$name <- gsub(coefs$name, pattern="Int\\)", replacement="Int")
coefs$name <- gsub(coefs$name, pattern="\\)", replacement="")
coefs$name <- gsub(coefs$name, pattern="I\\(", replacement="")
return(coefs)
}

plot.marginal <- function(sp, mod, par, par.name, type)
{
  nd <- data.frame(nd.mean[,names(nd.mean) != par],
                   par=seq(min(site.stacked[,par]), max(site.stacked[,par]), length.out=1000))
  names(nd) <- gsub(names(nd), pattern="par", replacement=par)
  marginal <- unmarked::predict(mod, type=type, newdata=nd)

  o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
    geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.5, fill="grey75") +
    geom_rug(data=site.stacked, aes(x=site.stacked[,par], y=0)) +
    ylim(0,1) +
    xlab(par.name) + ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +
    geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
  print(o)
  ggsave(o, filename=paste("../Figures/occupancy/", sp, "_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")
}


```

## load data
data:
- removed ME 2nd and 3rd surveys (only kept 1st survey to stay consistent with other states)
- removed minutes 7-10 of CT surveys to stay consistent with protocols from other states

```{r}
ewpw.data <- read.csv("../Data_processed/occupancy_detectionhistory_stacked_EWPW.csv")
cwwi.data <- read.csv("../Data_processed/occupancy_detectionhistory_stacked_CWWI.csv")
coni.data <- read.csv("../Data_processed/occupancy_detectionhistory_stacked_CONI.csv")

table(ewpw.data$MaxDetections) #includes a bunch from MA missing observation data
 # -Inf     0     1     2     3     4     5     6     7     8     9    10    50 
 #   97 12177  3752  1340   587   191    63    25     6     4     3     1     1
table(cwwi.data$MaxDetections) #includes 12 sites from ME missing observation data
 # -Inf     0     1     2     3     4     5     6     7     8     9 
 #   12 10810  1913  1213   894   414   134    36    10     6     2 
table(coni.data$MaxDetections)
 # -Inf     0     1     2     3     4     5     7     9    24 
 #   12 14363   716   210    58    14     2     1     1     1 

table(ewpw.data$state)
 #  CT   DE   FL   GA   MA   MD   ME   NC   NH   NY   PA   SC   VA   WV 
 # 657    9 2565  806 2109  231 3081 2486 1868 1603  587  886 1328   31

#select data here
data <- ewpw.data
# data <- cwwi.data
# data <- coni.data

# fix moon data
# ME: 1st Sunset (A), 2nd Sunset (B), 1st Moonrise (C), or 2nd Moonrise (D): this is more about start_time
# twilight = time between day and night when sun is below the horizon but rays still light up the sky
#1st sunset = civil twilight (sun has disappeared below the horizon); starts at sunset and ends when center of the sun is 6 degrees below the horizon
#2nd sunset = nautical twilight (Sun's disk is between 6 and 12 degrees below the horizon)
data$moon <- toupper(data$moon)
data$moon <- trimws(data$moon)
data[data$moon %in% c("A","B","C","D"),]$moon <- NA
data[data$moon %in% c("Y"),]$moon <- 1
data[data$moon %in% c("N"),]$moon <- 0
data[data$moon %in% c(""),]$moon <- NA

#remove rows where there is no detection information
data <- data[!is.na(data$MaxDetections),]
data <- data[data$MaxDetections != "-Inf",]

#reorder data by state and year
data <- data[order(data$state, data$year),]

#start with only data with all covariates (remove rows with NA in siteCovs)
summary(data)

#####
#remove data for detection models
# data <- data[!is.na(data$start_date_j),] #removes most of NH sites since they don't have month/day or start_time info
# data <- data[!is.na(data$sky),] #removes most of the MA sites since they don't have sky coded yet
# data <- data[!is.na(data$moon),]
# data <- data[!is.na(data$noise),]

#####
#remove data for occupancy only models
# #WV doesn't have any Lat/Long information
data <- data[!is.na(data$BCR),] #remove sites without BCR info
data <- data[!is.na(data$year),] #remove sites without year info

data <- data[!is.na(data$Route.Latitude),] #try this to see if latitude is a better predictor
data <- data[!is.na(data$Route.Longitude),]

data <- data[data$year != 2023,] #remove 2023 data since only from DE
# data <- data[data$year != 2008,] #remove for global global model

########
#fix classes for data
data$hour <- as.factor(data$hour)

data$BCR <- factor(data$BCR)
data$year <- factor(data$year)

data$noise <- as.numeric(data$noise)
data$moon <- factor(data$moon)

table(data$state)

plot(data$hour, data$TotalDetections)
plot(data$hour, data$MaxDetections)

nrow(data) #EWPW: 13,229 rows for detection models, 15,877 rows for occupancy models (12376 rows if remove those without latitude info)
summary(data)

sum(data$TotalDetections) #this should match number of 1s in the summary of the umf

```

## create umf for stacked single-season occupancy model
```{r}
sp.stacked <- data[,grep(names(data), pattern="min")]
sp.stacked[sp.stacked > 1 & !is.na(sp.stacked)] <- 1 #convert > 1 to 1 (presence/absence instead of counts)

#check detection histories
unique(sp.stacked)

site.stacked <- data[,!names(data) %in% grep(names(data), pattern="min", value=T)] #siteCovs

#add in transformations for variables
site.stacked$start_date_j2 <- scale(site.stacked$start_date_j)^2
site.stacked$start_time2 <- scale(site.stacked$start_time_s)^2

site.stacked$Route.Latitude2 <- scale(site.stacked$Route.Latitude)^2
site.stacked$Route.Longitude2 <- scale(site.stacked$Route.Longitude)^2

min.df <- matrix(rep(1:6, times=nrow(sp.stacked)), byrow=T, ncol=6) #add obsCovs, dimensions the same as y data

umf <- unmarkedFrameOccu(y=sp.stacked, siteCovs=site.stacked, obsCovs=list(min=min.df))
head(umf)

tiff(filename="../Figures/occupancy/EWPW_dethist_nsn_stacked.tiff", height=10, width=8, units="in", compression="lzw", res=400)
plot(umf)
dev.off()

summary(umf)

#y=1/y=0
15579/(31869 + 15579) #only 32.8% of occasions where EWPW were detected (with all NA data removed) #for the detection only model
14162/(62777 + 14162) #only 18.4% of occasions where EWPW were detected #for the occupancy only model
#XX% of occasions where EWPW were detected for the detection + occupancy model
```

## run detection models for stacked single season occupancy models
include random effect of route and state

7908 sites
Maximum number of observations per site: 6 
Mean number of observations per site: 6 
Sites with at least one detection: 3322 

Tabulation of y observations:
    0     1 
31869 15579 

```{r, echo=T}
null <- occu(~1 ~1, data=umf)
null.r <- occu(~(1|st_id) + (1|state) ~1, data=umf)
summary(null.r)

plogis(-1.58) #detection -> 17.1% detection
plogis(0.796) #occupancy -> 68.9% occupancy

#####
#detection models
#####
p.min <- occu(~scale(min) + (1|st_id) + (1|state) ~1, data=umf)

#
p.moon <- occu(~moon + (1|st_id) + (1|state) ~1, data=umf)
p.sky <- occu(~scale(sky) + (1|st_id) + (1|state) ~1, data=umf)
p.noise <- occu(~scale(noise) + (1|st_id) + (1|state) ~1, data=umf)

p.day <- occu(~scale(start_date_j) + (1|st_id) + (1|state) ~1, data=umf)
p.day2 <- occu(~scale(start_date_j) + I(scale(start_date_j) * scale(start_date_j)) + 
                 (1|st_id) + (1|state) ~1, data=umf)
p.month <- occu(~scale(start_month) + (1|st_id) + (1|state) ~1, data=umf)
p.month2 <- occu(~scale(start_month) + I(scale(start_month)*scale(start_month)) + 
                   (1|st_id) + (1|state) ~1, data=umf)
p.hour <- occu(~hour + (1|st_id) + (1|state) ~1, data=umf)
p.time <- occu(~scale(start_time_s) + (1|st_id) + (1|state) ~1, data=umf)
p.time2 <- occu(~scale(start_time_s) + scale(I(start_time_s^2)) + (1|st_id) + (1|state) ~1, data=umf)

p.time.c <- occu(~scale(start_time_c) + (1|st_id) + (1|state) ~1, data=umf)

p.timecat <- occu(~start_time_cat + (1|st_id) + (1|state) ~1, data=umf)

#######
p.fl <- fitList(null, null.r, p.min, #per minute within a survey
                p.moon, p.sky, p.noise, #site covariates
                p.day, p.day2, p.month, p.month2, p.time, p.time.c, p.time2, p.timecat, p.hour)

modSel(p.fl)
#           nPars      AIC   delta   AICwt cumltvWt
# p.noise       3 26347.59    0.00 1.0e+00     1.00
# p.day2        4 26582.24  234.65 1.1e-51     1.00
# p.month2      4 26607.29  259.70 4.0e-57     1.00 #
# p.moon        3 26619.65  272.06 8.4e-60     1.00
# p.month       3 26636.54  288.95 1.8e-63     1.00
# p.day         3 26637.27  289.68 1.3e-63     1.00 #
# p.min         3 26657.70  310.11 4.6e-68     1.00
# p.sky         3 26664.26  316.67 1.7e-69     1.00
# p.hour       20 26674.04  326.45 1.3e-71     1.00 #
# p.timecat     5 26687.70  340.11 1.4e-74     1.00
# p.time.c      3 26692.90  345.31 1.0e-75     1.00 #

# null.r        2 26695.64  348.05 2.6e-76     1.00
# p.time        3 26697.16  349.57 1.2e-76     1.00
# p.time2       4 26698.13  350.54 7.6e-77     1.00
# null          2 31686.44 5338.85 0.0e+00     1.00

p.global <- occu(~scale(min) + 
                   scale(noise) + moon + scale(sky) +
                   scale(start_date_j) + start_date_j2 +
                   scale(start_time_c) + 
                   (1|st_id) + (1|state) ~1, data=umf)
p.global2 <- occu(~scale(min) + 
                   scale(noise) + moon + scale(sky) +
                   scale(start_date_j) + I(scale(start_date_j)^2) +
                   scale(start_time_c) + 
                   (1|st_id) + (1|state) ~1, data=umf)
```
## coefficients and marginal plots for detection variables
include random effect of route and state
```{r, echo=T}
p.coefs <- vals.coefs(p.global) #choose model to extract coefficients from

c.plot <- ggplot(data=p.coefs[!p.coefs$name %in% "Int",], aes(x=name, y=est, group=p, shape=p)) +
  # c.plot <- ggplot(data=coefs[coefs$name != "INT",], aes(x=name, y=est)) +
  geom_point(size=3, position=position_dodge(width=0.5)) + 
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.5)) +
  # geom_errorbar(aes(ymin=est-se, ymax=est+se), width=0, lwd=1, position=position_dodge(width=0.5)) +
  geom_hline(aes(yintercept=0), lty="dotted") +
  xlab("parameter") + ylab("estimate") +
  coord_flip() +
  theme_bw(base_size=16) + theme(panel.grid=element_blank(), legend.position="none") + facet_wrap(.~p, scales="free", ncol=1)
c.plot

ggsave(c.plot, filename="../Figures/occupancy/EWPW_stacked_occupancy_det_coefficients.tiff", height=6, width=4, units="in", dpi=400, compression="lzw")

###########
#univariate model marginal plots
plotEffects(p.min, "det", covariate="min")
plotEffects(p.noise, "det", covariate="noise")
plotEffects(p.day2, "det", covariate="start_date_j")
plotEffects(p.month2, "det", covariate="start_month")
plotEffects(p.timecat, "det", covariate="start_time_cat")
plotEffects(p.hour, "det", covariate="hour")

###########
# #global model marginal plots for each covariate

plotEffects(p.global2, "det", covariate="start_date_j")
# 
# these won't work with p.global2, problem seems to be using I(date^2) notation
plotEffects(p.global, "det", covariate="min")
plotEffects(p.global, "det", covariate="noise")
plotEffects(p.global, "det", covariate="moon")
plotEffects(p.global, "det", covariate="sky")
# 
# ###########
# #manually plot marginals in ggplot2
nd.mean <- site.stacked %>% mutate_if(is.numeric, mean, na.rm=T)
nd.mean <- nd.mean[10,] #only take one row (plots depend on which site chosen)
nd.mean$min <- 1

ylabel <- data.frame(type=c("det","state"), label=c("detection","occupancy"))
#
#plot these with p.global2, need to pull out scale() from within I() to plot these
plot.marginal(sp="EWPW", p.global, "noise", "noise", type="det")
plot.marginal(sp="EWPW", p.global, "sky", "sky", type="det")
plot.marginal(sp="EWPW", p.global, "start_time_c", "start time", type="det")

plot.marginal(sp="EWPW", p.global2, "start_date_j", "Julian day", type="det") #need to plot separately with p.global since have scale within I()

#categorical
#for moon
# plot.marginal(sp="EWPW", p.global, "moon", "moon", type="det")

sp <- "EWPW"; par <- "moon"; par.name <- "moon visible"; type <- "det"
nd <- data.frame(nd.mean[,names(nd.mean) != par], par=0:1)
nd$par <- factor(nd$par, levels=c(0,1))
names(nd) <- gsub(names(nd), pattern="par", replacement=par)
marginal <- predict(p.global, type=type, newdata=nd)
o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.5, lwd=1) + geom_point(size=3) +
  ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +ylim(0,1) +
  xlab(par.name) +
  geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
print(o)
ggsave(o, filename=paste("../Figures/occupancy/", sp, "_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")


#for min
mod <- p.global; par <- "min"; par.name <- "minute"; type <- "det"
nd <- data.frame(nd.mean[,names(nd.mean) != par], par=1:6)
names(nd) <- gsub(names(nd), pattern="par", replacement=par)
marginal <- unmarked::predict(mod, type=type, newdata=nd)
o <- ggplot(marginal, aes(x=nd[,par], y=Predicted)) +
  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.5, fill="grey75") +
  ylab(paste("Predicted probability of", ylabel[ylabel$type == type,]$label)) +ylim(0,1) +
  xlab(par.name) + scale_x_continuous(breaks=1:6) +
  geom_line() + theme_bw(base_size=20) + theme(panel.grid=element_blank())
print(o)
ggsave(o, filename=paste("../Figures/occupancy/EWPW_stacked_occupancy_",type,"_", par,".tiff", sep=""), height=6, width=6, unit="in", dpi=300, compression="lzw")

```

## run occupancy models for stacked single season occupancy models
include random effect of route on occupancy, included state as a fixed effect

Without sites without lat/long info
14495 sites
Maximum number of observations per site: 6 
Mean number of observations per site: 5.31 
Sites with at least one detection: 4024 

Tabulation of y observations:
    0     1  <NA> 
62777 14162 10031 

```{r, echo=T}
#####
#occupancy models
#####

null <- occu(~1 ~ (1|st_id), data=umf)

#univariate
o.state <- occu(~1 ~ state + (1|st_id), data=umf)
o.bcr <- occu(~1 ~ BCR + (1|st_id), data=umf) #removing this so can run models with more sites and BCR does not increase fit by much
o.year <- occu(~1 ~ year + (1|st_id), data=umf)

#not enough data to run this, will run but SEs get weird 
o.lat <- occu(~1 ~ scale(Route.Latitude) + (1|st_id), data=umf)
o.lat2 <- occu(~1 ~ scale(Route.Latitude) + I(scale(Route.Latitude)* scale(Route.Latitude)) + (1|st_id), data=umf)
o.long <- occu(~1 ~ scale(Route.Longitude) + (1|st_id), data=umf)
o.long2 <- occu(~1 ~ scale(Route.Longitude) + I(scale(Route.Longitude)*scale(Route.Longitude)) + (1|st_id), data=umf)
# o.state.bcr.year.lat2 <- occu(~1 ~ state + BCR + year + scale(Route.Latitude) + I(scale(Route.Latitude)* scale(Route.Latitude))+ (1|st_id), data=umf)

o.fl <- fitList(o.lat, o.lat2, o.long, o.long2)
modSel(o.fl)
#         nPars      AIC  delta   AICwt cumltvWt
# o.lat2      4 34680.97   0.00 1.0e+00     1.00
# o.long2     4 34781.62 100.66 1.4e-22     1.00
# o.lat       3 34817.71 136.74 2.0e-30     1.00
# o.long      3 34843.17 162.21 6.0e-36     1.00

o.fl <- fitList(null,
                o.lat, o.lat2, o.long, o.long2,
                o.bcr, o.state, o.year)
modSel(o.fl)
#         nPars      AIC  delta   AICwt cumltvWt
# o.state    14 34595.55   0.00 1.0e+00     1.00
# o.bcr       8 34618.65  23.10 9.6e-06     1.00
# o.lat2      4 34680.97  85.42 2.8e-19     1.00
# o.long2     4 34781.62 186.07 3.9e-41     1.00
# o.lat       3 34817.71 222.16 5.7e-49     1.00
# o.year     17 34839.79 244.24 9.2e-54     1.00
# null        2 34841.26 245.71 4.4e-54     1.00
# o.long      3 34843.17 247.62 1.7e-54     1.00

#covariate combinations
o.state.bcr <- occu(~1 ~ state + BCR + (1|st_id), data=umf)
o.state.year <- occu(~1 ~ state + year + (1|st_id), data=umf)

o.state.bcr.year <- occu(~1 ~ state + BCR + year + (1|st_id), data=umf)

o.global <- occu(~1 ~ scale(Route.Latitude) + Route.Latitude2 + 
                   scale(Route.Longitude) + Route.Longitude2 + 
                   BCR + year + (1|st_id), data=umf)
o.global1 <- occu(~1 ~ scale(Route.Latitude) + I(scale(Route.Latitude)^2) + 
                   scale(Route.Longitude) + Route.Longitude2 + 
                   BCR + year + (1|st_id), data=umf)
o.global2 <- occu(~1 ~ scale(Route.Latitude) + Route.Latitude2 + 
                   scale(Route.Longitude) + I(scale(Route.Longitude)^2) + 
                   BCR + year + (1|st_id), data=umf)

```

## plot occupancy only model figures
```{r}
plotEffects(o.lat2, "state", covariate="Route.Latitude")
plotEffects(o.long2, "state", covariate="Route.Longitude")
# plotEffects(o.state, "state", covariate="state")
# plotEffects(o.bcr, "state", covariate="BCR")
# 
# plotEffects(o.state.year, "state", covariate="state")
# plotEffects(o.state.year, "state", covariate="year")
# 
# plotEffects(o.state.year, "state", covariate="state")
# plotEffects(o.state.year, "state", covariate="year")

plotEffects(o.state.bcr.year, "state", covariate="year")
plotEffects(o.state.bcr.year, "state", covariate="state")
plotEffects(o.state.bcr.year, "state", covariate="BCR")

# plotEffects(o.state.bcr.year.lat2, "state","Route.Latitude")

####
#plot occupancy per state per BCR per year
####
# nd <- data.frame(state=rep(unique(data$state), times=length(unique(data$year))), 
#                  year=rep(unique(data$year), each=length(unique(data$state))),
#                  st_id="GA--270056") #have to provide this for predict to work

mod <- o.state.bcr.year
nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
nd <- nd %>% #take first row of each group
  group_by(state, BCR, year) %>%
  arrange(st_id) %>%
  filter(row_number()==1)

marginal <- predict(mod, type="state", newdata=nd)

m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))

p <- ggplot(m.data, aes(x=year, y=Predicted, col=BCR, fill=BCR)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, pch=21, col="black", position=position_dodge(width=0.25)) +
  # geom_smooth(method="lm", col="grey50", lty="dashed") +
  scale_fill_brewer(palette="Set2") + scale_color_brewer(palette="Set2") +
  xlab("state") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
p + facet_wrap(~state, ncol=1)
ggsave(filename = "../Figures/occupancy/EWPW_stacked_occupancy_psi_state-bcr-year.tiff", height=30, width=16, units="in", dpi=400, compression="lzw")

p + facet_wrap(~state, ncol=2, dir="v")
ggsave(filename = "../Figures/occupancy/EWPW_stacked_occupancy_psi_state-bcr-year2.tiff", height=20, width=20, units="in", dpi=400, compression="lzw")

######
#plot values for each state
######
nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
nd <- nd %>% #take first row of each group
  group_by(state) %>%
  arrange(st_id) %>%
  filter(row_number()==1)
nd$year <- factor(2018)
marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))

ggplot(m.data, aes(x=state, y=Predicted)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, col="black", position=position_dodge(width=0.25)) +
  xlab("state") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occ_psi_state.tiff", height=6, width=8, units="in", dpi=400, compression="lzw")

#######
#plot values for each BCR
#######
nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
# nd <- nd %>% #take first row of each group
#   group_by(BCR) %>%
#   arrange(st_id) %>%
#   filter(row_number()==1)

nd <- nd %>% #take first row of each group
  group_by(BCR) %>%
  arrange(st_id) %>%
  filter(year==2018) %>%
  filter(row_number()==1)

marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))
m.data$BCR <- factor(m.data$BCR, levels=rev(c(31,27,29,28,30,13,14)))

ggplot(m.data, aes(x=BCR, y=Predicted)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, col="black", position=position_dodge(width=0.25)) +
  xlab("BCR") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occ_psi_BCR.tiff", height=6, width=8, units="in", dpi=400, compression="lzw")

#can't really plot values for each year since each state was only samples sporadically

##############
#force use of o.global so uniform with CONI and CWWI
##############
nd.mean <- site.stacked %>% mutate_if(is.numeric, mean, na.rm=T)
nd.mean <- nd.mean[10,] #only take one row since all rows have the same numeric values (but st_id matters)

ylabel <- data.frame(type=c("det","state"), label=c("detection","occupancy"))

plot.marginal(sp=sp, mod=o.global1, par="Route.Latitude", par.name="Latitude", type="state")
plot.marginal(sp=sp, mod=o.global2, par="Route.Longitude", par.name="Longitude", type="state")

require(sf)
AF_states <- st_read(dsn="../../GIS", layer="AF_states")
BCR.ST <- read_sf(dsn="../../GIS", layer="BCR-state") #area in mi^2
BCR <- BCR.ST %>% group_by(BCR) %>% dplyr::summarise(AreaSqMi = sum(AreaSqMi, na.rm=T), nstates=n())

mod <- o.global

#melt state boundaries within BCR
#calculate centroid of each BCR, use to predict
#need to provide st_id for each state
nd.centroid <- st_centroid(BCR) %>% st_coordinates()
st_id <- unique(site.stacked[,c("BCR","st_id","state")]) %>%
  group_by(BCR) %>%
  arrange(st_id) %>%
  filter(row_number()==1)

unique(site.stacked[,c("BCR","st_id","state")]) %>% filter(state == "ME") %>% filter(BCR==14) #use ME instead of CT or NY for BCR 14
st_id[st_id$BCR == 14,]$state <- "ME"
st_id[st_id$BCR == 14,]$st_id <- "ME--Durham"

nd <- cbind(BCR %>% st_drop_geometry, nd.centroid)
names(nd) <- c("BCR","AreaSqMi","nstate","Route.Longitude","Route.Latitude")

nd <- merge(nd, st_id, by=c("BCR"), all.x=T)
nd$Route.Latitude2 <- scale(nd$Route.Latitude)^2
nd$Route.Longitude2 <- scale(nd$Route.Longitude)^2
nd$BCR <- factor(nd$BCR)
nd$year <- factor(2018)

marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
# m.data <- m.data[order(m.data$state),]
# m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))
m.data$BCR <- factor(m.data$BCR, levels=c(31,27,29,28,30,13,14))

ggplot(m.data, aes(x=BCR, y=Predicted)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, col="black", position=position_dodge(width=0.25)) +
  xlab("BCR") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occupancy_psi_BCR.tiff", height=6, width=8, units="in", dpi=400, compression="lzw")

###########
#BCR + year plot
nd <- cbind(BCR %>% st_drop_geometry, nd.centroid)
names(nd) <- c("BCR","AreaSqMi","nstate","Route.Longitude","Route.Latitude")

nd <- merge(nd, st_id, by=c("BCR"), all.x=T)
nd$Route.Latitude2 <- scale(nd$Route.Latitude)^2
nd$Route.Longitude2 <- scale(nd$Route.Longitude)^2
nd$BCR <- factor(nd$BCR)

u.year <- unique(site.stacked[,c("BCR","year")])
nd <- merge(nd, u.year, by="BCR", all.x=T)
nd$Route.Latitude2 <- scale(nd$Route.Latitude)^2
nd$Route.Longitude2 <- scale(nd$Route.Longitude)^2

marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
# m.data <- m.data[order(m.data$state),]
# m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))
# m.data$BCR <- factor(m.data$BCR, levels=rev(c(31,27,29,28,30,13,14)))

ggplot(m.data, aes(x=year, y=Predicted, col=BCR)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.75)) +
  geom_point(size=3, position=position_dodge(width=0.75)) +
  scale_color_brewer(palette="Set2") + 
  xlab("Year") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occupancy_psi_BCR-year.tiff", height=6, width=16, units="in", dpi=400, compression="lzw")



```

## run det + psi models for stacked single season occupancy models
include random effect of route and state on detection
include random effect of route on occupancy

7480 sites
Maximum number of observations per site: 6 
Mean number of observations per site: 6 
Sites with at least one detection: 2349 

Tabulation of y observations:
    0     1 
35792  9088


More important to run with more sites

15278 sites
Maximum number of observations per site: 6 
Mean number of observations per site: 6 
Sites with at least one detection: 3994 

Tabulation of y observations:
    0     1  <NA> 
76432 15225    11 

*this is a much smaller dataset since need covariate info and BCR info*

```{r, echo=T}
#####
#detection + occupancy models
#####
null <- occu(~1 ~1, data=umf)
null.r <- occu(~(1|st_id) ~(1|st_id), data=umf)
  
# p.global <- occu(~scale(min) + scale(noise) + 
#                    scale(start_date_j) + start_date_j2 + #I(scale(start_date_j)*scale(start_date_j)) +
#                    moon + #scale(sky) + #removed since almost no effect
#                    (1|st_id) + (1|state) ~1, data=umf)

o.global <- occu(~1 ~ state + BCR + year + (1|st_id), data=umf)

#add variables to detection starting with o.global
p.null.o.global <- occu(~(1|st_id) ~ state + BCR + year + (1|st_id), data=umf)
p.min.o.global <- occu(~scale(min) + (1|st_id) ~ state + BCR + year + (1|st_id), data=umf)


# NaN for SEs
# global <- occu(~scale(min) + scale(noise) + 
#                    scale(start_date_j) + start_date_j2 +
#                    scale(moon) + #scale(sky) + #remove since almost no effect
#                    (1|st_id) + (1|state)
#                ~BCR + year + #state + 
#                  (1|st_id), data=umf)

p.global.o.year <- occu(~scale(min) + scale(noise) + 
                   scale(start_date_j) + start_date_j2 +
                   scale(moon) + #scale(sky) + #remove since almost no effect
                   (1|st_id) + (1|state)
               ~year + (1|st_id), data=umf)
p.global.o.BCR <- occu(~scale(min) + scale(noise) + 
                   scale(start_date_j) + start_date_j2 +
                   scale(moon) + #scale(sky) + #remove since almost no effect
                   (1|st_id) + (1|state)
               ~BCR + (1|st_id), data=umf)
p.global.o.state <- occu(~scale(min) + scale(noise) + 
                   scale(start_date_j) + start_date_j2 +
                   scale(moon) + #scale(sky) + #remove since almost no effect
                   (1|st_id) + (1|state)
               ~state + (1|st_id), data=umf)

fl <- fitList(null, null.r,
              # p.global, 
              o.global, 
              p.null.o.global, p.min.o.global
              # p.global.o.year, p.global.o.BCR, p.global.o.state
              )
modSel(fl)

#                  nPars      AIC   delta    AICwt cumltvWt
# p.global.o.BCR      11 21399.21    0.00  1.0e+00     1.00
# p.global.o.state    14 21501.17  101.96  7.2e-23     1.00
# p.global.o.year     20 21557.89  158.68  3.5e-35     1.00
# null.r               2 21710.17  310.96  3.0e-68     1.00
# o.global            26 21936.21  537.00 2.5e-117     1.00
# p.global             7 23081.90 1682.69  0.0e+00     1.00
# null                 2 27644.53 6245.32  0.0e+00     1.00
```


## plot marginals for stacked single season occupancy models
```{r, echo=T}

plotEffects(p.global.o.year, "state", covariate="year")
plotEffects(p.global.o.state, "state", covariate="state")
plotEffects(p.global.o.BCR, "state", covariate="BCR")

plotEffects(p.min.o.global, "state", covariate="state")
plotEffects(p.min.o.global, "state", covariate="year")
plotEffects(p.min.o.global, "state", covariate="BCR")

plotEffects(p.null.o.global, "state", covariate="year")

mod <- p.min.o.global

####
#plot occupancy per state per BCR per year
####
# nd <- data.frame(state=rep(unique(data$state), times=length(unique(data$year))), 
#                  year=rep(unique(data$year), each=length(unique(data$state))),
#                  st_id="GA--270056") #have to provide this for predict to work

nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
nd <- nd %>% #take first row of each group
  group_by(state, BCR, year) %>%
  arrange(st_id) %>%
  filter(row_number()==1)

marginal <- predict(mod, type="state", newdata=nd)

m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))

p <- ggplot(m.data, aes(x=year, y=Predicted, col=BCR, fill=BCR)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, pch=21, col="black", position=position_dodge(width=0.25)) +
  # geom_smooth(method="lm", col="grey50", lty="dashed") +
  scale_fill_brewer(palette="Set2") + scale_color_brewer(palette="Set2") +
  xlab("state") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
p + facet_wrap(~state, ncol=1)
ggsave(filename = "../Figures/occupancy/EWPW_stacked_occupancy_psi_state-bcr-year.tiff", height=30, width=16, units="in", dpi=400, compression="lzw")

p + facet_wrap(~state, ncol=2, dir="v")
ggsave(filename = "../Figures/occupancy/EWPW_stacked_occupancy_psi_state-bcr-year2.tiff", height=20, width=20, units="in", dpi=400, compression="lzw")

######
#plot values for each state
######
nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
nd <- nd %>% #take first row of each group
  group_by(state, BCR) %>%
  arrange(st_id) %>%
  filter(row_number()==1)
nd$year <- as.factor(2022)
marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))

ggplot(m.data, aes(x=state, y=Predicted, col=BCR)) +
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=5, position=position_dodge(width=0.25)) +
  xlab("state") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) + scale_color_brewer(palette = "Set2") + 
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occ_psi_state.tiff", height=6, width=8, units="in", dpi=400, compression="lzw")

#######
#plot values for each BCR
#######
nd <- unique(site.stacked[,c("state","BCR","year","st_id")])
nd <- nd %>% #take first row of each group
  group_by(state, BCR) %>%
  arrange(st_id) %>%
  filter(row_number()==1)
nd$year <- as.factor(2022)
marginal <- predict(mod, type="state", newdata=nd)
m.data <- cbind(nd, marginal)
m.data <- m.data[order(m.data$year, m.data$state),]
m.data$state <- factor(m.data$state, levels=rev(c("FL","GA","SC","NC","VA","WV","MD","DE","NJ","PA","NY","CT","MA","VT","NH","ME")))
m.data$BCR <- factor(m.data$BCR, levels=rev(c(31,27,29,28,30,13,14)))

ggplot(m.data, aes(x=BCR, y=Predicted, group=state, col=state)) +
  scale_color_
  geom_hline(aes(yintercept=0), lty="dashed", col="grey50") + geom_hline(aes(yintercept=1), lty="dashed", col="grey50") +
  geom_errorbar(aes(ymin=lower, ymax=upper), width=0.1, lwd=1, position=position_dodge(width=0.25)) +
  geom_point(size=4, position=position_dodge(width=0.25)) +
  xlab("BCR") + ylab(paste("Predicted probability of occupancy")) + ylim(c(0,1)) +
  theme_bw(base_size=20) + theme(panel.grid=element_blank())
ggsave(filename="../Figures/occupancy/EWPW_stacked_occ_psi_BCR.tiff", height=6, width=8, units="in", dpi=400, compression="lzw")
```
